add column maintenances table

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\AppController;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CarController;
 use App\Http\Controllers\CustomerController;
+use App\Http\Controllers\MaintenanceController;
 use App\Http\Controllers\OperatorController;
 use App\Http\Controllers\PaymentController;
 use App\Http\Controllers\RentalController;
@@ -28,34 +29,43 @@ Route::get('/admin/dashboard', [AppController::class, 'dashboard'])->name('admin
 
 Route::prefix('admin')->name('admin.car.')->middleware('auth')->group(function () {
     Route::get('/cars', [CarController::class, 'index'])->name('index');
-    Route::post('/cars', [CarController::class, 'store'])->name('store');
-    Route::post('/cars/{id}', [CarController::class, 'update'])->name('update');
-    Route::delete('/cars/{id}', [CarController::class, 'destroy'])->name('destroy');
+    Route::post('/cars', [CarController::class, 'store'])->name('store')->middleware('admin');
+    Route::post('/cars/{id}', [CarController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/cars/{id}', [CarController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::prefix('admin')->name('admin.customer.')->middleware('auth')->group(function () {
     Route::get('/customer', [CustomerController::class, 'index'])->name('index');
-    Route::post('/customer', [CustomerController::class, 'store'])->name('store');
-    Route::put('/customer/{id}', [CustomerController::class, 'update'])->name('update');
-    Route::delete('/customer/{id}', [CustomerController::class, 'destroy'])->name('destroy');
+    Route::post('/customer', [CustomerController::class, 'store'])->name('store')->middleware('admin');
+    Route::put('/customer/{id}', [CustomerController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/customer/{id}', [CustomerController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 
 Route::prefix('admin')->name('admin.transaction.')->middleware('auth')->group(function () {
     Route::get('/transaction', [TransactionController::class, 'index'])->name('index');
-    Route::post('/transaction', [TransactionController::class, 'store'])->name('store');
-    Route::put('/transaction/{uuid}', [TransactionController::class, 'update'])->name('update');
-    Route::delete('/transaction/{uuid}', [TransactionController::class, 'destroy'])->name('destroy');
+    Route::post('/transaction', [TransactionController::class, 'store'])->name('store')->middleware('admin');
+    Route::put('/transaction/{uuid}', [TransactionController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/transaction/{uuid}', [TransactionController::class, 'destroy'])->name('destroy')->middleware('admin');
+    Route::post('/returned/{uuid}', [TransactionController::class, 'returnedCar'])->name('return.car');
+    Route::post('/lunas/{uuid}', [TransactionController::class, 'lunasPayment'])->name('lunas.payment');
 });
 
-Route::prefix('admin')->name('admin.operator.')->middleware('auth')->group(function () {
+Route::prefix('admin')->name('admin.operator.')->middleware('auth', 'admin')->group(function () {
     Route::get('/operator', [OperatorController::class, 'index'])->name('index');
     Route::post('/operator', [OperatorController::class, 'store'])->name('store');
     Route::put('/operator/{uuid}', [OperatorController::class, 'update'])->name('update');
     Route::delete('/operator/{uuid}', [OperatorController::class, 'destroy'])->name('destroy');
+    Route::post('/operator/{uuid}', [OperatorController::class, 'changePassword'])->name('change.password');
 });
 
-
+Route::prefix('admin')->name('admin.maintenance.')->middleware('auth', 'admin')->group(function () {
+    Route::get('/maintenance', [MaintenanceController::class, 'index'])->name('index');
+    Route::post('/maintenance', [MaintenanceController::class, 'store'])->name('store');
+    Route::put('/maintenance/{uuid}', [MaintenanceController::class, 'update'])->name('update');
+    Route::delete('/maintenance/{uuid}', [MaintenanceController::class, 'destroy'])->name('destroy');
+});
+    
 Route::prefix('admin')->name('admin.')->group(function () {
     Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
     Route::post('/login', [AuthController::class, 'loginStore'])->name('loginStore')->middleware('guest');
